tests/wpt/tests/fetch/metadata/generated/element-link-icon.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-link-icon.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <meta name="timeout" content="long">
  <title>HTTP headers on request for HTML "link" element with rel="icon"</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  const params = {
    body: `
      <svg xmlns="http://www.w3.org/2000/svg" width="123" height="123">
        <rect fill="lime" width="123" height="123"/>
      </svg>
    `,
    mime: 'image/svg+xml'
  };

  /**
   * The `link` element supports a `load` event. That event would reliably
   * indicate that the browser had received the request. Multiple major
   * browsers do not implement the event, however, so in order to promote the
   * visibility of this test, a less efficient polling-based detection
   * mechanism is used.
   *
   * https://bugzilla.mozilla.org/show_bug.cgi?id=1638188
   * https://bugs.chromium.org/p/chromium/issues/detail?id=1083034
   */
  function induceRequest(t, url, attributes) {
    const link = document.createElement('link');
    link.setAttribute('rel', 'icon');
    link.setAttribute('href', url);

    for (const [ name, value ] of Object.entries(attributes)) {
      link.setAttribute(name, value);
    }

    document.head.appendChild(link);
    t.add_cleanup(() => link.remove());
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsOrigin'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
      });
  }, 'sec-fetch-site - Same origin no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsCrossSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Cross-site no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsSameSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same site no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite', 'httpsOrigin'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsOrigin', 'httpsSameSite', 'httpsOrigin'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsCrossSite', 'httpsOrigin'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Cross-Site -> Same Origin no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsCrossSite', 'httpsSameSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Cross-Site -> Same-Site no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsCrossSite', 'httpsCrossSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Cross-Site -> Cross-Site no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsOrigin', 'httpsOrigin'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
      });
  }, 'sec-fetch-site - Same-Origin -> Same Origin no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsOrigin', 'httpsSameSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same-Origin -> Same-Site no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsSameSite', 'httpsOrigin'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same-Site -> Same Origin no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsSameSite', 'httpsSameSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same-Site -> Same-Site no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpsSameSite', 'httpsCrossSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Same-Site -> Cross-Site no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, [], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['no-cors']);
      });
  }, 'sec-fetch-mode no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, [], params),
      {"crossorigin": ""}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['cors']);
      });
  }, 'sec-fetch-mode attributes: crossorigin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, [], params),
      {"crossorigin": "anonymous"}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['cors']);
      });
  }, 'sec-fetch-mode attributes: crossorigin=anonymous');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, [], params),
      {"crossorigin": "use-credentials"}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['cors']);
      });
  }, 'sec-fetch-mode attributes: crossorigin=use-credentials');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, [], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_array_equals(headers['sec-fetch-dest'], ['empty']);
      });
  }, 'sec-fetch-dest no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, [], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
      });
  }, 'sec-fetch-user no attributes');
  </script>
  </body>
</html>
